agent-evals/pyproject.toml
[project]
name = "agent-evals"
version = "0.1.0"
description = "Tests index formats, structures, and doc transformations to find which produces the best agent outcomes"
requires-python = ">=3.11"
dependencies = [
    "agent-index",
    "pydantic>=2.0",
    "litellm>=1.40",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "numpy>=1.24",
    "tiktoken>=0.5",
    "python-dotenv>=1.0",
    "datasets>=2.14",
    "huggingface-hub>=0.20",
    "rich>=13.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "sse-starlette>=1.6",
    "jinja2>=3.0",
    "httpx>=0.24",
]

[project.scripts]
agent-evals = "agent_evals.cli:main"
observatory = "agent_evals.cli:dashboard_main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.uv.sources]
agent-index = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_evals"]
